fixed error with bad pull

+++ server/db/migrations/07-create-estimation.js
@@ -24,7 +24,7 @@ module.exports = {
         },
       },
       value: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.NUMERIC,
       },
       createdAt: {
         allowNull: false,
